Refactor: Move bg details to _app
