+++ pages/components/AppFooterContent.module.scss
@@ -1,0 +1,5 @@
+.footer {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
