+++ pages/components/AppFooterContent.tsx
@@ -1,10 +1,10 @@
-import styles from "./AppFooter.module.scss";
+import styles from "./AppFooterContent.module.scss";
 
 export default function AppFooter() {
   return (
-    <footer className={styles.footer}>
+    <div className={styles.footer}>
       <div className="copyright">&copy; NightlyHerb 2022</div>
       <nav>Social</nav>
-    </footer>
+    </div>
   );
 }
