--- pages/components/AppHeader.module.scss
@@ -1,15 +1,0 @@
-@use "./main-banner.scss" as *;
-
-:where(.header) {
-  position: sticky;
-  top: 0;
-}
-
-:where(.header) {
-  @include main-banner;
-}
-
-:where(.header nav) {
-  display: flex;
-  align-items: center;
-}
