+++ pages/components/AppHeaderContent.module.scss
@@ -1,0 +1,5 @@
+.header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
