+++ pages/components/AppHeaderContent.tsx
@@ -1,15 +1,15 @@
 import React from "react";
 
-import styles from "./AppHeader.module.scss";
+import styles from "./AppHeaderContent.module.scss";
 
 export default function AppHeader() {
   return (
-    <header className={styles.header}>
+    <div className={styles.header}>
       <div className="title">NightlyHerb Blog</div>
       <nav>
         <a>Posts</a>
         <a>Tags</a>
       </nav>
-    </header>
+    </div>
   );
 }
