--- pages/components/main-content.scss
@@ -1,6 +1,0 @@
-@use "~open-color" as *;
-
-@mixin main-content {
-  background-color: $oc-teal-1;
-  color: $oc-gray-9;
-}
